Remove interactors docs
