+++ website/docs/4-architecture.md
