+++ website/netlify.toml
@@ -1,0 +1,12 @@
+[[redirects]]
+  from = "/bigtest/interactors/*"
+  to = "https://frontside.com/interactors/:splat"
+
+[[redirects]]
+  from = "/bigtest/docs/interactors/*"
+  to = "https://frontside.com/interactors/docs/:splat"
+
+[[redirects]]
+  from = "/bigtest/docs/platform/*"
+  to = "/bigtest/docs/:splat"
+
