+++ website/sidebars.js
@@ -1,19 +1,8 @@
 module.exports = {
-  interactors: [
-    'interactors/quick-start',
-    'interactors/predefined-interactors',
-    'interactors/locators-filters-actions',
-    'interactors/assertions',
-    'interactors/matchers',
-    'interactors/write-your-own',
-    'interactors/integrations',
-    'interactors/storybook',
-    'interactors/existing-interactors',
-  ],
   platform: [
-    'platform/installation',
-    'platform/writing-your-first-test',
-    'platform/running-tests',
-    'platform/architecture',
+    'installation',
+    'writing-your-first-test',
+    'running-tests',
+    'architecture',
   ],
 };
