--- website/src/pages/docs.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { Redirect } from '@docusaurus/router';
-import useBaseUrl from '@docusaurus/useBaseUrl';
-
-function DocsRedirect() {
-  return <Redirect to={useBaseUrl('/docs/interactors')} />;
-}
-
-export default DocsRedirect;
